Added all of the window, key and mouse events

+++ Comet/src/Comet/Core/MouseCodes.h
@@ -4,6 +4,6 @@
 
 namespace comet {
 
-	using MouseCode = uint16_t;
+	using MouseCode = uint32_t;
 
 }
